Theo/requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
presidio-analyzer>=2.2.0
presidio-anonymizer>=2.2.0
spacy>=3.7.0
openai>=1.0.0
python-dotenv>=1.0.0
